Bug 732114 - part 3: add dumpDB into AndroidBrowserRepositoryDataAccessor.
