+++ src/main/java/org/mozilla/gecko/sync/repositories/android/RepoUtils.java
@@ -187,4 +187,58 @@ public class RepoUtils {
     
     return a.equals(b);
   }
+
+  private static String fixedWidth(int width, String s) {
+    if (s == null) {
+      return spaces(width);
+    }
+    int length = s.length();
+    if (width == length) {
+      return s;
+    }
+    if (width > length) {
+      return s + spaces(width - length);
+    }
+    return s.substring(0, width);
+  }
+
+  private static String spaces(int i) {
+    return "                                     ".substring(0, i);
+  }
+
+  public static void dumpCursor(Cursor cur) {
+    int originalPosition = cur.getPosition();
+    try {
+      String[] columnNames = cur.getColumnNames();
+      int columnCount      = cur.getColumnCount();
+
+      // 12 chars each column.
+      for (int i = 0; i < columnCount; ++i) {
+        System.out.print(fixedWidth(12, columnNames[i]) + " | ");
+      }
+      System.out.println("");
+      for (int i = 0; i < columnCount; ++i) {
+        System.out.print("------------" + " | ");
+      }
+      System.out.println("");
+      if (!cur.moveToFirst()) {
+        System.out.println("EMPTY");
+        return;
+      }
+
+      cur.moveToFirst();
+      while (cur.moveToNext()) {
+        for (int i = 0; i < columnCount; ++i) {
+          System.out.print(fixedWidth(12, cur.getString(i)) + " | ");
+        }
+        System.out.println("");
+      }
+      for (int i = 0; i < columnCount; ++i) {
+        System.out.print("---------------");
+      }
+      System.out.println("");
+    } finally {
+      cur.moveToPosition(originalPosition);
+    }
+  }
 }
